Added missing functionality and bug fixes
Implemented parameter to choose whether you want to receive the fulltext or not when querying documents and plenary protocols

Implemented the filters for Deskriptor, Sachgebiet, Dokumententyp and Titel

Added some very rudimentary logging

Added some util functions and separated the entity classes
