+++ README.md
@@ -1,5 +1,7 @@
 # Bundestag-API
-A wrapper for the official Bundestag (German Federal Parliament) API in Python.
+A wrapper for the official Bundestag (German Federal Parliament) API in Python. It aim is to make querying the API a little easier in Python and relieve you from writing a lot of boilerplate code. 
+
+It currently doesn't encompass a 100% of the lastest API parameters (see To-Do section) but aims to do so shortly.
 
 The official information on the API can be found here: [Bundestag.de](https://dip.bundestag.de/%C3%BCber-dip/hilfe/api)
 
@@ -20,14 +22,14 @@ $ pip install -r requirements.txt
 ```
 
 ## Setup
-The API requires a key to authenticate requests. Personal key can be requested from the [Bundestag administration](mailto:parlamentsdokumentation@bundestag.de). However a general API key has been published that is valid until May 31st 2023. This key is automatically used until that date when no other key is supplied by the user.
+The API requires a key to authenticate requests. Personal key can be requested from the [Bundestag administration](mailto:parlamentsdokumentation@bundestag.de). However a general API key has been published that is valid until May 31st 2024. This key is automatically used until that date when no other key is supplied by the user.
 
 ## Usage
-To save your API key create a connection-object, that you can then pass to the search functions. It will save you time, should you have to change API keys at a later date. If you don't supply an API key, the official API key will be used until 31st of May 2023. 
+To save your API key create a connection-object, that you can then pass to the search functions. It will save you time, should you have to change API keys at a later date. If you don't supply an API key, the official API key will be used until 31st of May 2024. 
 ```
 import bundestag_api
 bta = bundestag_api.btaConnection() #if you want to use your own API key, supply it via "apikey="XXX")
-data = bta.query(resource="vorgang")
+data = bta.search_document()
 for d in data:
     print(d["drucksachetyp"]+": "+d["titel"])
 ```
@@ -45,7 +47,7 @@ Get one or more activities by their ID
 ```
 bta.get_activity(btid)
 ```
-Search for documents by specifying parameters for start and end date or institution. Important: The standard number of documents returned are 100. If more are desired, the "num" parameter must be set.
+Search for documents by specifying parameters for start and end date or institution. Important: The standard number of entities returned are 100. If more are desired, the "num" parameter must be set.
 ```
 bta.search_activity()
 ```
@@ -54,18 +56,18 @@ Get one or more documents by their ID
 ```
 bta.get_document(btid)
 ```
-Search for documents by specifying parameters for start and end date or institution. Important: The standard number of documents returned are 100. If more are desired, the "num" parameter must be set.
+Search for documents by specifying parameters for start and end date or institution. Important: The standard number of entities returned are 100. If more are desired, the "num" parameter must be set.
 ```
 bta.search_document()
 ```
-The API differentiates between documents and documents incuding full-text of the document. The functions always query the full-text resource.
+"fulltext=True" can be passed as parameter to retrieve the full text of the document (if available). It defaults to False.
 
 ### Persons ("Person")
 Get one or more persons by their ID
 ```
 bta.get_person(btid)
 ```
-Search for persons by specifying parameters for start and end date or institution. Important: The standard number of documents returned are 100. If more are desired, the "num" parameter must be set.
+Search for persons by specifying parameters for start and end date or institution. Important: The standard number of entities returned are 100. If more are desired, the "num" parameter must be set.
 ```
 bta.search_person()
 ```
@@ -74,18 +76,18 @@ Get one or more plenary protocols by their ID
 ```
 bta.get_plenaryprotocol(btid)
 ```
-Search for plenary protocols by specifying parameters for start and end date or institution. Important: The standard number of documents returned are 100. If more are desired, the "num" parameter must be set.
+Search for plenary protocols by specifying parameters for start and end date or institution. Important: The standard number of entities returned are 100. If more are desired, the "num" parameter must be set.
 ```
 bta.search_plenaryprotocol()
 ```
-The API differentiates between plenary protocols and plenary protocols incuding full-text of the protocol. The functions always query the full-text resource.
+"fulltext=True" can be passed as parameter to retrieve the full text of the plenary protocols (if available). It defaults to False.
 
 ### Procedures ("Vorgang")
 Get one or more procedures by their ID
 ```
 bta.get_procedure(btid)
 ```
-Search for procedures by specifying parameters for start and end date or institution. Important: The standard number of documents returned are 100. If more are desired, the "num" parameter must be set.
+Search for procedures by specifying parameters for start and end date or institution. Important: The standard number of entities returned are 100. If more are desired, the "num" parameter must be set.
 ```
 bta.search_procedure()
 ```
@@ -94,11 +96,15 @@ Get one or more procedure positions by their ID
 ```
 bta.get_procedure(btid)
 ```
-Search for procedure positions by specifying parameters for start and end date or institution. Important: The standard number of documents returned are 100. If more are desired, the "num" parameter must be set.
+Search for procedure positions by specifying parameters for start and end date or institution. Important: The standard number of entities returned are 100. If more are desired, the "num" parameter must be set.
 ```
 bta.search_procedure()
 ```
 
 ## ToDo's
+- Implement filters for GESTA-Number, Beratungsstand, Fundstelle, Initiative, Ressort (federführend), Verkündungsblatt_Kürzel, Vorgangstyp, Vorgangstyp-Notation
+- Implement retries before failure
+- Implement sufficient unit tests
+- Implement more extensive logging
 - Parallelize requests for larger queries
 - Extend Class methods
