+++ bundestag_api/__init__.py
@@ -3,5 +3,5 @@
 Python wrapper for the official Bundestag API
 """
 
-from .bta_wrapper import btaConnection, Person, Role, Drucksache, \
-    Aktivitaet, Vorgang, Vorgangsposition, Plenarprotokoll
+from .bta_wrapper import btaConnection
+from .models import, Person, Role, Drucksache, Aktivitaet, Vorgang, Vorgangsposition, Plenarprotokoll
